refactor: Rename `…s/aws/package/compile/events/iotFleetProvisioning.js`
