+++ lib/plugins/aws/package/compile/events/iot-fleet-provisioning.js
