+++ lib/plugins/index.js
@@ -45,7 +45,7 @@ module.exports = [
   require('./aws/package/compile/events/alexa-skill.js'),
   require('./aws/package/compile/events/alexa-smart-home.js'),
   require('./aws/package/compile/events/iot.js'),
-  require('./aws/package/compile/events/iotFleetProvisioning.js'),
+  require('./aws/package/compile/events/iot-fleet-provisioning.js'),
   require('./aws/package/compile/events/cloud-watch-event.js'),
   require('./aws/package/compile/events/cloud-watch-log.js'),
   require('./aws/package/compile/events/cognito-user-pool.js'),
