+++ test/unit/lib/plugins/aws/package/compile/events/iot-fleet-provisioning/index.test.js
